Tweak

+++ db/schema.rb
@@ -151,7 +151,7 @@ ActiveRecord::Schema.define(version: 20_190_719_233_448) do
     t.string "subdivision_name", limit: 100, default: "", null: false, comment: "Second-level administrative division (e.g. county/district/division/province/etc.) of this location if available"
     t.string "city_name", limit: 100, default: "", null: false, comment: "City (or equivalent) of this location if available"
     t.bigint "osm_id", comment: "OpenStreetMap ID for traceability. May change at any time."
-    t.integer "locationiq_id", comment: "Data provider API ID for traceability."
+    t.bigint "locationiq_id", comment: "Data provider API ID for traceability."
     t.decimal "lat", precision: 10, scale: 6, comment: "The latitude of this location if available"
     t.decimal "lng", precision: 10, scale: 6, comment: "The longitude of this location if available"
     t.datetime "created_at", null: false
